(tae): small refactor 2
